requirements.txt
# Django and DRF
Django==4.2.7
djangorestframework==3.14.0

# Authentication and OAuth
django-allauth==0.57.0

# Database
psycopg2-binary>=2.9.10

# Environment variables
python-dotenv==1.0.0

# WSGI and ASGI
gunicorn==21.2.0
daphne==4.0.0

# Static files
whitenoise==6.6.0

# Image handling
Pillow>=11.0.0

# Development
django-debug-toolbar==4.2.0
